+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - Updated action to Node v20.
 - Updated dependencies.
   - `@actions/github` v5 -> v6
+  - `conventional-commits-parser` v3 -> v6